Added memoization and did performance optimizations
